+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
+## next
+* Add `modalClassName` prop to `Modal`
+
 ## 2.0.0
 * Update to react 15.6.2